+++ src/App.tsx
@@ -1,297 +1,10 @@
 import { useEffect, useState } from "react"
-import { Plus, Minus, ArrowLeft, Pencil, Check, Copy, ShoppingBag } from "lucide-react"
 import { api } from "./api/apiClient"
-import { slugify, brl, buildQrGrid } from "./helpers"
-import { Eyebrow } from "./components/Eyebrow"
-import { ScreenShell } from "./components/ScreenShell"
+import { slugify } from "./helpers"
 import { StatusScreen } from "./components/StatusScreen"
-
-const QR_GRID = buildQrGrid()
-const AVATAR_CLASSES = ["bg-teal", "bg-stamp", "bg-mustard"]
-
-// ---------------- Screen 1: pick seller ----------------
-function SellerScreen({ sellers, selectedId, setSelectedId, onCreateSeller, creating, onNext }) {
-    return (
-        <ScreenShell>
-            <div className="flex-1 flex flex-col px-4 sm:px-6 pt-4 pb-6 overflow-y-auto">
-                <Eyebrow step={1} label="Vendedor" />
-                <h1 className="mt-1 mb-6 leading-tight font-display font-bold text-ink text-[26px]">
-                    Quem tá vendendo hoje?
-                </h1>
-
-                <div className="flex flex-col gap-2">
-                    {sellers.length === 0 && (
-                        <p className="font-body text-inkFaint text-[13px]">
-                            Nenhum vendedor cadastrado ainda no CAAD-ERP.
-                        </p>
-                    )}
-                    {sellers.map((s, i) => {
-                        const isSel = selectedId === s.id
-                        return (
-                            <button
-                                key={s.id}
-                                onClick={() => setSelectedId(s.id)}
-                                className={`flex items-center gap-3 px-4 py-3 rounded-2xl text-left transition-all border-[1.5px] border-solid ${isSel
-                                    ? "bg-tealSoft border-teal"
-                                    : "bg-card border-paperLine"
-                                    }`}
-                            >
-                                <div className={`rounded-full flex items-center justify-center shrink-0 w-10 h-10 text-white font-display font-bold text-base ${AVATAR_CLASSES[i % AVATAR_CLASSES.length]}`}>
-                                    {s.name.charAt(0).toUpperCase()}
-                                </div>
-                                <span className="font-body text-ink font-semibold text-[15px]">
-                                    {s.name}
-                                </span>
-                                {isSel && (
-                                    <span className="ml-auto">
-                                        <Check size={18} className="text-teal" />
-                                    </span>
-                                )}
-                            </button>
-                        )
-                    })}
-
-                    <button
-                        disabled={creating}
-                        onClick={onCreateSeller}
-                        className={`flex items-center gap-3 px-4 py-3 rounded-2xl border-[1.5px] border-dashed border-inkFaint bg-transparent disabled:opacity-60`}
-                    >
-                        <div className="rounded-full flex items-center justify-center shrink-0 w-10 h-10 border-[1.5px] border-dashed border-inkFaint">
-                            <Plus size={16} className="text-inkSoft" />
-                        </div>
-                        <span className="font-body text-inkSoft font-semibold text-[15px]">
-                            {creating ? "Cadastrando..." : "Novo vendedor"}
-                        </span>
-                    </button>
-                </div>
-            </div>
-
-            <div className="px-4 sm:px-6 pb-7 pt-3 shrink-0 border-t border-dashed border-paperLine">
-                <button
-                    disabled={!selectedId}
-                    onClick={onNext}
-                    className="w-full py-3.5 rounded-2xl transition-opacity text-white font-display font-bold text-[15px] bg-ink disabled:bg-inkFaint disabled:opacity-60"
-                >
-                    Começar venda
-                </button>
-            </div>
-        </ScreenShell>
-    )
-}
-
-// ---------------- Screen 2: cart ----------------
-function CartScreen({ seller, products, stock, qty, setQty, onBack, onClose }) {
-    const items = products.map((p) => ({ ...p, qty: qty[p.id] || 0 })).filter((p) => p.qty > 0)
-    const total = items.reduce((s, p) => s + p.qty * p.price, 0)
-
-    const availableFor = (id) => stock[id]
-    const inc = (id) =>
-        setQty((q) => {
-            const current = q[id] || 0
-            const available = availableFor(id)
-            if (available !== undefined && current >= available) return q
-            return { ...q, [id]: current + 1 }
-        })
-    const dec = (id) => setQty((q) => ({ ...q, [id]: Math.max(0, (q[id] || 0) - 1) }))
-
-    return (
-        <ScreenShell>
-            <div className="px-4 sm:px-6 pt-4 shrink-0">
-                <Eyebrow step={2} label="Carrinho" />
-                <div className="flex items-center gap-2 mb-4">
-                    <button onClick={onBack} className="p-1 -ml-1">
-                        <ArrowLeft size={20} className="text-ink" />
-                    </button>
-                    <h1 className="font-display text-ink text-xl font-bold">
-                        Venda de {seller?.name}
-                    </h1>
-                    <button onClick={onBack} className="ml-auto p-1">
-                        <Pencil size={14} className="text-inkFaint" />
-                    </button>
-                </div>
-            </div>
-
-            <div className="flex-1 overflow-y-auto px-4 sm:px-6">
-                <p className="text-xs uppercase tracking-widest mb-2 font-body text-inkFaint font-semibold">
-                    Toque para adicionar
-                </p>
-                <div className="grid grid-cols-3 gap-2 mb-5">
-                    {products.map((p) => {
-                        const available = availableFor(p.id)
-                        const soldOut = available !== undefined && available <= 0
-                        return (
-                            <button
-                                key={p.id}
-                                onClick={() => inc(p.id)}
-                                disabled={soldOut}
-                                className="flex flex-col items-center justify-center gap-1 rounded-2xl py-3 relative bg-card border border-solid border-paperLine disabled:opacity-45"
-                            >
-                                {qty[p.id] > 0 && (
-                                    <span className="absolute -top-1.5 -right-1.5 rounded-full flex items-center justify-center w-5 h-5 bg-teal text-white font-mono text-[11px] font-semibold">
-                                        {qty[p.id]}
-                                    </span>
-                                )}
-                                <span className="text-[20px]">{p.emoji}</span>
-                                <span className="font-body text-ink text-xs font-semibold">
-                                    {p.name}
-                                </span>
-                                <span className="font-mono text-inkSoft text-[11px]">
-                                    {soldOut ? "Esgotado" : brl(p.price)}
-                                </span>
-                            </button>
-                        )
-                    })}
-                </div>
-
-                <div className="pt-3 border-t border-dashed border-paperLine">
-                    {items.length === 0 ? (
-                        <div className="flex flex-col items-center text-center py-10 gap-2">
-                            <ShoppingBag size={28} className="text-inkFaint" />
-                            <p className="font-body text-inkFaint text-[13px]">
-                                Nenhum item ainda.
-                                <br />
-                                Toque em um produto acima pra começar.
-                            </p>
-                        </div>
-                    ) : (
-                        items.map((p) => (
-                            <div key={p.id} className="flex items-center gap-2 py-2">
-                                <span className="text-base">{p.emoji}</span>
-                                <span className="flex-1 font-body text-ink text-sm font-medium">
-                                    {p.name}
-                                </span>
-                                <div className="flex items-center gap-2 rounded-full px-1 bg-paper border border-solid border-paperLine">
-                                    <button onClick={() => dec(p.id)} className="p-1">
-                                        <Minus size={12} className="text-inkSoft" />
-                                    </button>
-                                    <span className="font-mono text-xs text-ink min-w-3.5 text-center">
-                                        {p.qty}
-                                    </span>
-                                    <button onClick={() => inc(p.id)} className="p-1">
-                                        <Plus size={12} className="text-inkSoft" />
-                                    </button>
-                                </div>
-                                <span className="font-mono text-ink text-[13px] min-w-15.5 text-right">
-                                    {brl(p.qty * p.price)}
-                                </span>
-                            </div>
-                        ))
-                    )}
-                </div>
-            </div>
-
-            <div className="px-4 sm:px-6 pt-3 pb-7 shrink-0 border-t-[1.5px] border-dashed border-inkFaint">
-                <div className="flex items-baseline justify-between mb-3">
-                    <span className="font-body text-inkSoft text-[13px] font-semibold">
-                        Total
-                    </span>
-                    <span className="font-mono text-ink text-2xl font-semibold">
-                        {brl(total)}
-                    </span>
-                </div>
-                <button
-                    disabled={total === 0}
-                    onClick={onClose}
-                    className="w-full py-3.5 rounded-2xl text-white font-display font-bold text-[15px] bg-teal disabled:bg-inkFaint disabled:opacity-60"
-                >
-                    Fechar venda
-                </button>
-            </div>
-        </ScreenShell>
-    )
-}
-
-// ---------------- Screen 3: PIX QR ----------------
-function PixScreen({ total, status, error, onConfirm, onNewSale }) {
-    const [copied, setCopied] = useState(false)
-    const confirmed = status === "confirmed"
-    const confirming = status === "confirming"
-
-    return (
-        <div
-            className="w-full min-h-svh flex flex-col items-center justify-center px-4 sm:px-5 bg-ink pt-[max(1rem,env(safe-area-inset-top))] pb-[max(1rem,env(safe-area-inset-bottom))] pl-[max(1rem,env(safe-area-inset-left))] pr-[max(1rem,env(safe-area-inset-right))]"
-        >
-            <Eyebrow step={3} label="Pagamento" />
-            <div className="relative w-full max-w-xs mt-2">
-                {confirmed && (
-                    <div className="absolute z-10 -top-[14px] -right-[6px] w-[84px] h-[84px] rounded-full border-[3px] border-solid border-stamp flex flex-col items-center justify-center -rotate-12 bg-stamp/[0.06]">
-                        <span className="font-display text-stamp font-bold text-[15px] tracking-wider">
-                            PAGO
-                        </span>
-                        <Check size={14} className="text-stamp" />
-                    </div>
-                )}
-
-                <div className="bg-card pt-5">
-                    <div className="flex flex-col items-center px-4 sm:px-6 pb-7">
-                        <span className={`mb-1 font-body font-bold text-xs tracking-[0.14em] ${confirmed ? "text-teal" : "text-mustard"}`}>
-                            {confirmed ? "PAGAMENTO CONFIRMADO" : "AGUARDANDO PAGAMENTO"}
-                            {!confirmed && <span className="animate-pulse ml-1">●</span>}
-                        </span>
-
-                        <div
-                            className="grid my-4"
-                            style={{ gridTemplateColumns: `repeat(${QR_SIZE}, 15px)`, gridTemplateRows: `repeat(${QR_SIZE}, 15px)` }}
-                        >
-                            {QR_GRID.flatMap((row, r) =>
-                                row.map((cell, c) => (
-                                    <div
-                                        key={r + "-" + c}
-                                        className={`w-3.75 h-3.75 ${cell ? "bg-ink" : "bg-transparent"}`}
-                                    />
-                                ))
-                            )}
-                        </div>
-
-                        <span className="font-mono text-ink text-[26px] font-semibold">
-                            {brl(total)}
-                        </span>
-                        <span className="font-body text-inkFaint text-xs mt-0.5">
-                            Pix • código de pagamento
-                        </span>
-
-                        <button
-                            onClick={() => {
-                                setCopied(true)
-                                setTimeout(() => setCopied(false), 1500)
-                            }}
-                            className="flex items-center gap-2 mt-4 px-3 py-2 rounded-xl w-full justify-center bg-paper border border-solid border-paperLine"
-                        >
-                            <span className="font-mono text-[11px] text-inkSoft">
-                                {copied ? "Copiado!" : "pix •••• 8f2a-91cd"}
-                            </span>
-                            <Copy size={13} className="text-inkSoft" />
-                        </button>
-                    </div>
-                </div>
-            </div>
-
-            <div className="w-full max-w-xs mt-6">
-                {error && (
-                    <p className="mb-3 text-center font-body text-stampSoft text-xs">
-                        {error}
-                    </p>
-                )}
-                {!confirmed ? (
-                    <button
-                        onClick={onConfirm}
-                        disabled={confirming}
-                        className="w-full py-3.5 rounded-2xl bg-teal text-white font-display font-bold text-sm disabled:opacity-70"
-                    >
-                        {confirming ? "Registrando venda..." : "Já recebi o pagamento"}
-                    </button>
-                ) : (
-                    <button
-                        onClick={onNewSale}
-                        className="w-full py-3.5 rounded-2xl bg-transparent text-white border-[1.5px] border-solid border-[#ffffff55] font-display font-bold text-sm"
-                    >
-                        Nova venda
-                    </button>
-                )}
-            </div>
-        </div>
-    )
-}
+import { SellerScreen } from "./screens/SellerScreen"
+import { PixScreen } from "./screens/PixScreen"
+import { CartScreen } from "./screens/CartScreen"
 
 export default function App() {
     const [screen, setScreen] = useState("seller") // seller | cart | pix
